Fix circular import.
