+++ workbench/test/test_scenarios.py
@@ -10,7 +10,7 @@ from django.core.urlresolvers import reverse
 from xblock.test.tools import assert_equals
 
 from workbench import scenarios
-from workbench.runtime import reset_global_state
+from workbench.runtime_util import reset_global_state
 
 
 @ddt.ddt
